requirements.txt
# Core dependencies for DOCX document formatting
python-docx>=0.8.11
requests>=2.31.0
PyYAML>=6.0

# For overnight LLM processing (optional)
# Requires Ollama running locally at http://localhost:11434
# Install with: pip install requests (already included above)

# Development and testing (optional)
pytest>=7.0.0